Move EncryptedCast class to VirtualColumnTest
