--- tests/etc/EncryptedCast.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace Stancl\VirtualColumn\Tests\Etc;
-
-use Illuminate\Support\Facades\Crypt;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
-
-class EncryptedCast implements CastsAttributes
-{
-    /**
-     * Cast the given value.
-     *
-     * @param  array<string, mixed>  $attributes
-     */
-    public function get(Model $model, string $key, mixed $value, array $attributes): mixed
-    {
-        return Crypt::decryptString($value);
-    }
-
-    /**
-     * Prepare the given value for storage.
-     *
-     * @param  array<string, mixed>  $attributes
-     */
-    public function set(Model $model, string $key, mixed $value, array $attributes): mixed
-    {
-        return Crypt::encryptString($value);
-    }
-}
